Added esx_cctv
Added a fully customizable Closed-Circuit Television System with a complete RageUI built-in which will expand if you add your own cameras, there is no code to be written it is utilizing a simple `for k,v in pairs` function. Also there is more options such as Jobs which can access the CCTV Menu, the Bannercolor and Bannername.